requirements.txt
streamlit>=1.36.0
requests>=2.32.3
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
pandas>=2.2.2
python-dotenv>=1.0.1
tldextract>=5.1.2
phonenumbers>=8.13.45
trafilatura>=1.12.2
tenacity>=8.3.0
pydantic>=2.7.4
openai>=1.37.0
xlsxwriter>=3.2.0
